refactor(GCS+gRPC): reusable `SetContent` for cords
I am going to need `SetContent(ChecksummedData&, absl::Cord)` from more
than one `.cc` file.
